+++ CHANGELOG.md
@@ -813,6 +813,8 @@ For instructions on upgrading to newer versions, visit
 
 * \#2119 Allow `Criteria#update_all` to accept atomic ops and normal sets.
 
+* \#2118 Don't strip any precision during `DateTime` -> `Time` conversions.
+
 * \#2117 Ensure embeds one relations have callbacks run when using nested
   attributes.